refactor: Move repositories extensions to common module
