+++ infrastructure-common/src/main/kotlin/dsl/Repositories.kt
@@ -1,10 +1,20 @@
 package com.redmadrobot.build.dsl
 
+import org.gradle.api.Project
 import org.gradle.api.artifacts.dsl.RepositoryHandler
 import org.gradle.api.artifacts.repositories.MavenArtifactRepository
 import org.gradle.api.artifacts.repositories.PasswordCredentials
 import org.gradle.kotlin.dsl.credentials
 import org.gradle.kotlin.dsl.maven
+
+/**
+ * Checks that credentials for the repository with given [name] are existing.
+ *
+ * See: [Handling credentials](https://docs.gradle.org/current/userguide/declaring_repositories.html#sec:handling_credentials)
+ */
+public fun Project.credentialsExist(name: String): Boolean {
+    return hasProperty("${name}Username") && hasProperty("${name}Password")
+}
 
 /**
  * Adds RMR Nexus repo with name "rmrNexus".
@@ -51,50 +61,6 @@ public fun RepositoryHandler.githubPackages(
     return maven("https://maven.pkg.github.com/$repoName") {
         name = "githubPackages"
         credentials(PasswordCredentials::class)
-        configure()
-    }
-}
-
-/**
- * Adds Sonatype OSSRH repository with name "ossrh".
- * Requires credentials to be specified in project properties `ossrhUsername` and `ossrhPassword`.
- *
- * See:
- * [Handling Credentials](https://docs.gradle.org/current/userguide/declaring_repositories.html#sec:handling_credentials)
- * [OSSRH Guide](https://central.sonatype.org/pages/ossrh-guide.html)
- *
- * Example:
- * ```
- * repositories {
- *     ossrh()
- * }
- * ```
- * @see ossrhSnapshots
- */
-public fun RepositoryHandler.ossrh(configure: MavenArtifactRepository.() -> Unit = {}): MavenArtifactRepository {
-    return maven("https://oss.sonatype.org/service/local/staging/deploy/maven2/") {
-        name = "ossrh"
-        credentials(PasswordCredentials::class)
-        configure()
-    }
-}
-
-/**
- * Adds Sonatype OSSRH Snapshots repository with name "ossrhSnapshots".
- *
- * Example:
- * ```
- * repositories {
- *     ossrhSnapshots()
- * }
- * ```
- * @see ossrh
- */
-public fun RepositoryHandler.ossrhSnapshots(
-    configure: MavenArtifactRepository.() -> Unit = {},
-): MavenArtifactRepository {
-    return maven("https://oss.sonatype.org/content/repositories/snapshots/") {
-        name = "ossrhSnapshots"
         configure()
     }
 }
